+++ R/unlabel.R
@@ -1,13 +1,12 @@
-#' @title Convert labelled data (frames) into normal classes
+#' @title Convert labelled vectors into normal classes
 #' @name unlabel
 #'
-#' @description This function converts a vector or data frame, which was imported with any of
-#'                \code{haven}'s read functions and contains \code{\link[haven]{labelled}} class vectors or
-#'                a single vector of type \code{labelled} into an generic data
-#'                frame format, which means that simply all \code{\link[haven]{labelled}} class
-#'                attributes will be removed, so all vectors / variables will most
+#' @description This function converts \code{\link[haven]{labelled}} class vectors
+#'                into a generic data format, which means that simply all \code{\link[haven]{labelled}}
+#'                class attributes will be removed, so all vectors / variables will most
 #'                likely become \code{\link{atomic}}. Additionally, \code{tbl_df} and
-#'                \code{tbl} class attributes will be removed from data frames. See 'Note'.
+#'                \code{tbl} class attributes will be removed from data frames, and
+#'                a \code{\link{lbl_df}} class attribute will be added. See 'Note'.
 #'
 #' @seealso \href{http://www.strengejacke.de/sjPlot/datainit/}{sjPlot manual: data initialization}
 #'
@@ -23,6 +22,8 @@
 #'         \code{tbl} class attributes may cause difficulties when indexing columns
 #'         like \code{data.frame[, colnr]} - only \code{data.frame[[colnr]]} seems
 #'         to be safe when accessing data frame columns from within function calls.
+#'         \cr \cr
+#'         Data frames will be converted into labelled data frames (see \code{\link{lbl_df}}).
 #'
 #' @importFrom utils txtProgressBar setTxtProgressBar
 #' @export
@@ -48,7 +49,7 @@ unlabel <- function(x) {
     }
     close(pb)
     # remove redundant class attributes
-    class(x) <- "data.frame"
+    class(x) <- c("lbl_df", "data.frame")
   } else {
     # remove labelled class
     if (is_labelled(x)) x <- unclass(x)