+++ NEWS.md
@@ -5,6 +5,7 @@
 * `rec_pattern` to generate recode patterns for the `rec` function.
 * `drop_labels` to drop labels from values with zero-counts.
 * `str_contains` to check whether a string contains another string pattern.
+* `r2` to compute R-squared values for linear (mixed) models.
 * `lbl_df` to create a labelled data frame, and related S3-generic print method for `lbl_df` objects.
 
 ## Changes to functions